refactor(storage): extract sqlite_store run lifecycle to _history_runs
A5 of the sqlite_store decomposition - the highest-blast-radius
extraction, intentionally last in the sqlite track. Moves 15 methods
that drive every analysis-run lifecycle event:

- create_run / finish_run / update_run_planned_count /
  increment_run_processed / increment_run_applied / update_run_status
  / recover_stale_runs / update_run_heartbeat / update_run_current_step
  (write path - 27+ call sites)
- get_run / list_recent_runs / find_runs_for_scope / stats /
  count_pending_review_runs / list_recent_events (read aggregates -
  every dashboard + /history view)

Each function takes the store as 'hs' and uses _lock + _connect().
_ensure_run_columns and all DDL stays in SQLiteHistoryStore.init()
untouched.

SQLiteHistoryStore keeps thin delegating wrappers on every name.
sqlite_store.py: 2089 -> 1497 LOC (-592, -28%). Cumulative across
A1+A2+A3+A4+A5: 3535 -> 1497 LOC (-58%) over five PRs.
